fix: move birthday easter egg to App.tsx so it shows on every route
Previously in HomePage.tsx, which is skipped when ConditionalHomePage
auto-redirects to Quick mode. Now renders globally via App.tsx.
